requirements.txt
altgraph==0.10.2
amqp==1.4.9
ansible==2.3.0.0
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==1.1.2
#asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
#bcrypt==3.1.4
bdist-mpkg==0.5.0
billiard==3.3.0.23
cached-property==1.3.1
celery==3.1.20
celery-with-redis==3.0
certifi==2017.11.5
cffi==1.11.2
click==6.7
constantly==15.1.0
coverage==4.4.1
#cryptography==2.1.4
csvImporter==0.1.3.2
Cython==0.27.3
DBUtils==1.2
dj-static==0.0.6
Django==1.11.7
django-adaptors==0.2.5
django-celery==3.2.2
django-celery-beat==1.1.0
django-celery-results==1.0.1
django-filter==1.1.0
django-nose==1.4.5
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
#docker==2.7.0
#docker-compose==1.18.0
#docker-pycreds==0.2.1
#dockerpty==0.4.1
#docopt==0.6.2
#docutils==0.14
#enum34==1.1.6
#ephem==3.7.6.0
fakeredis==0.9.0
Flask==0.12.2
flower==0.9.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
google-api-python-client==1.6.4
httplib2==0.10.3
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
kombu==3.0.37
lxml==4.1.1
macholib==1.5.1
MarkupSafe==1.0
mongoengine==0.15.0
mrjob==0.6.1
MySQL-python
mysql-replication==0.17
nose==1.3.7
paramiko==2.4.0
pbr==3.1.1
pika==0.9.14
pluggy==0.6.0
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
py2app==0.7.3
pycparser==2.18
pycrypto==2.6.1
PyDispatcher==2.0.5
Pygments==2.2.0
pymongo==3.6.0
PyMySQL==0.8.0
PyNaCl==1.2.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
pytest==3.4.0
python-dateutil==2.6.1
pytz==2013.7
PyYAML==3.12
redis==2.10.6
requests==2.18.4
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
static3==0.7.0
supervisor==3.3.4
tornado==4.5.3
urllib3==1.22
virtualenv==15.1.0
websocket-client==0.46.0
xlrd==1.1.0
